+++ Cargo.toml
@@ -4,7 +4,7 @@ panic = 'unwind'
 [workspace]
 members = [
 	'node',
-	'pallets/template',
+	'pallets/simple-staking',
 	'runtime/statemint',
 	'runtime/statemine',
 ]
