+++ pallets/simple-staking/src/mock.rs
@@ -1,4 +1,4 @@
-use crate as pallet_template;
+use crate as simple_staking;
 use sp_core::H256;
 use frame_support::parameter_types;
 use sp_runtime::{
@@ -17,7 +17,7 @@ frame_support::construct_runtime!(
 		UncheckedExtrinsic = UncheckedExtrinsic,
 	{
 		System: frame_system::{Pallet, Call, Config, Storage, Event<T>},
-		TemplateModule: pallet_template::{Pallet, Call, Storage, Event<T>},
+		SimpleStaking: simple_staking::{Pallet, Call, Storage, Event<T>},
 	}
 );
 
@@ -51,7 +51,7 @@ impl system::Config for Test {
 	type SS58Prefix = SS58Prefix;
 }
 
-impl pallet_template::Config for Test {
+impl pallet_simple_staking::Config for Test {
 	type Event = Event;
 }
